dom/xbl/test/test_bug944407.xul

<?xml version="1.0"?>
<?xml-stylesheet type="text/css" href="chrome://global/skin"?>
<?xml-stylesheet type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"?>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=944407
-->
<window title="Mozilla Bug 944407"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml">
  <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=944407"
     target="_blank">Mozilla Bug 944407</a>
  </body>

  <!-- test code goes here -->
  <script type="application/javascript">
  <![CDATA[

  /** Test for XBL bindings with script disabled. **/
  SimpleTest.waitForExplicitFinish();
  const Cu = Components.utils;
  Cu.import('resource://gre/modules/Services.jsm');

  function go() {

    // Disable javascript, and load the frame.
    function loadFrame() {
      ok(!Services.prefs.getBoolPref('javascript.enabled'), "Javascript should be disabled");
      $('ifr').setAttribute('src', 'http://mochi.test:8888/tests/dom/xbl/test/file_bug944407.html');
    }
    SpecialPowers.pushPrefEnv({ set: [['javascript.enabled', false]] }, loadFrame);
  }

  function finish() {
    SimpleTest.finish();
  }

  addLoadEvent(go);
  ]]>
  </script>
  <iframe id='ifr' />
</window>
